edutech_backend/requirements.txt
Django==4.2.7
pymongo==4.6.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.1.2
